sequelize 모델 수정, router, contoller 구조
