+++ server/routes/user/kakao.js
@@ -1,5 +1,5 @@
 var express = require('express');
-var controller = require('../controller/login');
+var controller = require('../controller/user/login');
 const router = express.Router();
 
 router.get('/getLogin/kakao', controller.viewKakaoLogin);
